Add collision and food systems
Snake will now be reset to the middle of the arena if it touches
the edges. Food can will now be moved and placed at a random
location when the snake touches it.

Pointing amethyst dependency at a temporary fork that solves my
freezing problem as mentioned here https://github.com/amethyst/amethyst/issues/729